Move tag specs back to htmlTags.ts and externalize strings
